requirements.dev.txt
google-cloud-storage==1.35.1
pytest==6.2.1
selenium==3.141.0
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tqdm==4.55.0
webdriver-manager==3.2.2